otra vez el req.body del post en el login lpm!!

+++ src/controllers/usersController.js
@@ -13,12 +13,12 @@ const controller2 = {
         return res.render('users', { users: users });
     },
 
-    detalleUsuario:(req, res) => {
+    detalleUsuario: (req, res) => {
         const id = req.params.id;
         const user = users.find(user => user.id == id);
         return res.render('userDetail', { user: user });
     },
-  
+
     login: (req, res) => {
         return res.render('userLogin');
     },
@@ -27,13 +27,34 @@ const controller2 = {
         let errors = validationResult(req);
 
         if (errors.isEmpty()) {
-            res.redirect('/');       
-    } else {
-        res.render('userLogin', {msgErrors: errors.mapped(), old: req.body});        
-    }         
+
+            const user = users.find(user => user.user_name == req.body.user_name);
+            console.log(user)
+
+            if (user) {
+                let result = bcryptjs.compareSync(req.body.password, user.password);
+                if (result) {
+                    delete user.password;
+                    req.session.user = user;
+                    if (req.body.remember) {
+                        res.cookie('user', user, { maxAge: 1000 * 60 * 60 * 24 * 7 });
+                    }
+                    return res.redirect('/users/perfil' + user.id);
+                } else {
+                    return res.render('userLogin', { msgErrors: { password: { msg: 'Contraseña incorrecta' } } });
+                }
+                
+
+            } else {
+                return res.render('userLogin', { msgErrors: { user_name: { msg: 'El usuario no existe' } } });
+            }
+
+        } else {
+            res.render('userLogin', { msgErrors: errors.mapped(), old: req.body });
+        }
     },
 
-    logout: function(req, res){
+    logout: function (req, res) {
         res.clearCookie('remember');
         req.session.destroy();
         res.redirect('/');
@@ -48,13 +69,17 @@ const controller2 = {
         let errors = validationResult(req);
         let user = req.body;
 
-        if (errors.isEmpty()) {   
+        if (errors.isEmpty()) {
 
-            let exist = users.find(user => user.email == req.body.email);
-            if (exist) {
-                return res.render('userRegister', {msgErrors: {email: {msg: 'El usuario ya existe'}}, old: req.body});
+            let existName = users.find(user => user.email == req.body.email);
+            if (existName) {
+                return res.render('userEdit', { msgErrors: { nombreUsuario: { msg: 'El usuario ya existe' } }, old: req.body });
             }
-            let { first_name, last_name, user_name, birth, password, email, admin} = user;
+            let existEmail = users.find(user => user.email == req.body.email);
+            if (existEmail) {
+                return res.render('userEdit', { msgErrors: { email: { msg: 'El usuario ya existe' } }, old: req.body });
+            }
+            let { first_name, last_name, user_name, birth, password, email, admin } = user;
             const newUser = {}
             newUser.id = users[users.length - 1].id + 1;
             newUser.first_name = first_name;
@@ -71,22 +96,22 @@ const controller2 = {
             fs.writeFileSync(usersFilePath, JSON.stringify(users));
 
             req.session.user = user;
-            
+
             if (req.body.remember) {
 
                 res.cookie('remember', user, { maxAge: 1000 * 60 * 60 * 24 * 30 });
             }
 
             res.redirect('/users/perfil/' + newUser.id);
-        
+
         } else {
-            res.render('userRegister', {msgErrors: errors.mapped(), old: req.body});     
+            res.render('userRegister', { msgErrors: errors.mapped(), old: req.body });
         }
-              
-},
+
+    },
 
     delete: (req, res) => {
-        
+
         const id = req.params.id;
         const user = users.find(users => users.id == id);
 
@@ -105,7 +130,7 @@ const controller2 = {
         const id = req.params.id;
         const user = users.find(user => user.id == id);
 
-        res.render('userEdit', { user: user });
+        res.render('userEdit', { user: user, id: id });;
     },
 
     editput: (req, res) => {
@@ -113,25 +138,35 @@ const controller2 = {
         const errors = validationResult(req);
         const id = req.params.id;
         const user = users.find(user => user.id == id);
-        const { first_name, last_name, user_name, birth, password, email, admin} = req.body;
-        
-        if (errors.isEmpty()) {     
+        delete user.user_name;
+        let existName = users.find(user => user.user_name == req.body.user_name);
+        if (existName) {
+            return res.render('userEdit', { msgErrors: { nombreUsuario: { msg: 'El usuario ya existe' } }, old: req.body, id: id });
+        }
+        delete user.email;
+        let existEmail = users.find(user => user.email == req.body.email);
+        if (existEmail) {
+            return res.render('userEdit', { msgErrors: { email: { msg: 'El usuario ya existe' } }, old: req.body, id: id });
+        }
+        const { first_name, last_name, user_name, birth, password, email, admin } = req.body;
 
-        user.id = user.id;
-        user.first_name = first_name;
-        user.last_name = last_name;
-        user.user_name = user_name;
-        user.birth = birth;
-        user.email = email;
-        user.password = password;
-        user.admin = parseInt(admin) ? parseInt(admin) : 0;
+        if (errors.isEmpty()) {
+
+            user.id = user.id;
+            user.first_name = first_name;
+            user.last_name = last_name;
+            user.user_name = user_name;
+            user.birth = birth;
+            user.email = email;
+            user.password = bcryptjs.hashSync(password, 10);
+            user.admin = parseInt(admin) ? parseInt(admin) : 0;
 
 
-        fs.writeFileSync(usersFilePath, JSON.stringify(users));
-        res.redirect('/');
-    } else {
-        res.render('userEdit', {msgErrors: errors.mapped(), user: req.body, id: id});     
-    }
+            fs.writeFileSync(usersFilePath, JSON.stringify(users));
+            res.redirect('/');
+        } else {
+            res.render('userEdit', { msgErrors: errors.mapped(), user: req.body, id: id });
+        }
 
     },
